Move geometry related files to src/geom
