+++ src/geom/hit.h
