+++ src/geom/ray.h
