+++ src/main.cc
@@ -1,13 +1,13 @@
 #include "image.h"
 #include "file.h"
 #include "log.h"
-#include "ray.h"
-#include "sphere.h"
 #include "camera.h"
 #include "random.h"
 #include "material.h"
 #include "thread_pool.h"
 #include "util/resource_finder.h"
+#include "geom/ray.h"
+#include "geom/sphere.h"
 #include "loader/obj_loader.h"
 #include "loader/image_loader.h"
